+++ admin-api/src/main/java/project/dailyge/app/core/notice/application/NoticeWriteUseCase.java
@@ -1,0 +1,7 @@
+package project.dailyge.app.core.notice.application;
+
+import project.dailyge.app.core.notice.application.command.NoticeCreateCommand;
+
+public interface NoticeWriteUseCase {
+    Long save(NoticeCreateCommand command);
+}
