+++ admin-api/src/main/java/project/dailyge/app/core/notice/presentation/response/NoticeCreateResponse.java
@@ -1,0 +1,9 @@
+package project.dailyge.app.core.notice.presentation.response;
+
+public record NoticeCreateResponse(Long noticeId) {
+
+    @Override
+    public String toString() {
+        return String.format("{\"noticeId\":\"%s\"}", noticeId);
+    }
+}
