+++ admin-api/src/main/java/project/dailyge/app/core/user/application/command/UserBlacklistCreateCommand.java
@@ -1,0 +1,5 @@
+package project.dailyge.app.core.user.application.command;
+
+public record UserBlacklistCreateCommand(String accessToken) {
+
+}
