--- admin-api/src/main/java/project/dailyge/app/notice/application/NoticeWriteUseCase.java
@@ -1,7 +1,0 @@
-package project.dailyge.app.notice.application;
-
-import project.dailyge.app.notice.application.command.NoticeCreateCommand;
-
-public interface NoticeWriteUseCase {
-    Long save(NoticeCreateCommand command);
-}
