--- admin-api/src/main/java/project/dailyge/app/notice/persistence/NoticeWriteDao.java
@@ -1,22 +1,0 @@
-package project.dailyge.app.notice.persistence;
-
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
-import project.dailyge.entity.notice.NoticeEntityWriteRepository;
-import project.dailyge.entity.notice.NoticeJpaEntity;
-
-@Repository
-@RequiredArgsConstructor
-public class NoticeWriteDao implements NoticeEntityWriteRepository {
-
-    private final EntityManager entityManager;
-
-    @Override
-    @Transactional
-    public Long save(final NoticeJpaEntity notice) {
-        entityManager.persist(notice);
-        return notice.getId();
-    }
-}
