--- admin-api/src/main/java/project/dailyge/app/notice/presentation/request/NoticeCreateRequest.java
@@ -1,35 +1,0 @@
-package project.dailyge.app.notice.presentation.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
-import project.dailyge.app.notice.application.command.NoticeCreateCommand;
-import project.dailyge.entity.notice.NoticeType;
-
-public record NoticeCreateRequest(
-    @NotNull
-    @NotBlank
-    @Size(max = 50)
-    String title,
-
-    @NotNull
-    @NotBlank
-    @Size(max = 5000)
-    String content,
-
-    @NotNull
-    NoticeType noticeType
-) {
-
-    public NoticeCreateCommand toCommand(final Long userId) {
-        return new NoticeCreateCommand(title, content, noticeType, userId);
-    }
-
-    @Override
-    public String toString() {
-        return String.format(
-            "{\"title\":\"%s\",\"content\":\"%s\",\"noticeType\":\"%s\"}",
-            title, content, noticeType
-        );
-    }
-}
