--- admin-api/src/main/java/project/dailyge/app/notice/presentation/response/NoticeCreateResponse.java
@@ -1,9 +1,0 @@
-package project.dailyge.app.notice.presentation.response;
-
-public record NoticeCreateResponse(Long noticeId) {
-
-    @Override
-    public String toString() {
-        return String.format("{\"noticeId\":\"%s\"}", noticeId);
-    }
-}
