--- admin-api/src/main/java/project/dailyge/app/user/application/command/UserBlacklistCreateCommand.java
@@ -1,5 +1,0 @@
-package project.dailyge.app.user.application.command;
-
-public record UserBlacklistCreateCommand(String accessToken) {
-
-}
